Problems by handeling both very small (rad) and very large (m) proj coordinates are solved by changing from int to float and by dropping some tests that seem unnecessary
git-svn-id: https://svn.met.no/diana/trunk@2144 0ccb8871-b436-41a3-ae7d-e3ac59973202

+++ src/diPlotModule.h
@@ -116,7 +116,7 @@ private:
   Area requestedarea;
   Area previousrequestedarea;
 
-  int plotw, ploth;     // width and height of plotwindow (pixels)
+  float plotw, ploth;     // width and height of plotwindow (pixels)
   bool resizezoom;      // should resizing zoom splot.area?
   bool showanno;        // show standard annotations
 
@@ -139,8 +139,8 @@ private:
   vector <LocationPlot*> locationPlots; // location (vcross,...) to be plotted
 
   // event-handling
-  int oldx, oldy,xmoved,ymoved;
-  int newx, newy;
+  float oldx, oldy,xmoved,ymoved;
+  float newx, newy;
   Area myArea;
   deque<Area> areaQ;
   int areaIndex;
